pyproject.toml
[project]
name = "fitted"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agate==1.9.1",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "attrs==25.4.0",
    "babel==2.17.0",
    "boto3==1.41.3",
    "botocore==1.41.3",
    "cachetools==6.2.2",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "colorama==0.4.6",
    "cryptography==46.0.3",
    "daff==1.4.2",
    "databricks-cli==0.18.0",
    "databricks-sdk==0.67.0",
    "databricks-sql-connector==4.1.3",
    "dbt-adapters==1.17.3",
    "dbt-common==1.36.0",
    "dbt-core==1.10.15",
    "dbt-databricks==1.11.0",
    "dbt-extractor==0.6.0",
    "dbt-protos==1.0.382",
    "dbt-semantic-interfaces==0.9.0",
    "dbt-spark==1.9.3",
    "deepdiff==8.6.1",
    "distro==1.9.0",
    "duckduckgo-search>=8.1.1",
    "et-xmlfile==2.0.0",
    "fastapi==0.121.2",
    "google-auth==2.43.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "importlib-metadata==8.7.0",
    "isodate==0.6.1",
    "jaraco-classes==3.4.0",
    "jaraco-context==6.0.1",
    "jaraco-functools==4.3.0",
    "jeepney==0.9.0",
    "jinja2==3.1.6",
    "jiter==0.12.0",
    "jmespath==1.0.1",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "keyring==25.6.0",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "leather==0.4.0",
    "lz4==4.4.5",
    "mangum==0.19.0",
    "markupsafe==3.0.3",
    "mashumaro==3.14",
    "mlflow>=3.9.0",
    "more-itertools==10.8.0",
    "msgpack==1.1.2",
    "networkx==3.5",
    "numpy==2.3.4",
    "oauthlib==3.3.1",
    "openai==2.8.0",
    "openpyxl==3.1.5",
    "orderly-set==5.5.0",
    "packaging>=23.1,<25.0",
    "pandas==2.2.3",
    "parsedatetime==2.6",
    "pathspec==0.12.1",
    "pinecone-client>=6.0.0",
    "protobuf==6.33.1",
    "pyarrow==22.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.23",
    "pydantic>=2.10.0,<2.12",
    "pydantic-core>=2.27.0,<2.41",
    "pyjwt==2.10.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-slugify==8.0.4",
    "pytimeparse==1.1.8",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "python-fasthtml>=0.12.0",
    "requests==2.32.5",
    "rpds-py==0.28.0",
    "rsa==4.9.1",
    "s3transfer==0.15.0",
    "secretstorage==3.4.1",
    "sentence-transformers>=5.2.2",
    "six==1.17.0",
    "sniffio==1.3.1",
    "snowplow-tracker==1.1.0",
    "sqlparams==6.2.0",
    "sqlparse==0.5.3",
    "starlette>=0.41.0",
    "tabulate==0.9.0",
    "text-unidecode==1.3",
    "thrift==0.20.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websockets==15.0.1",
    "zipp==3.23.0",
    "psycopg[binary]>=3.1.0",
    "psycopg-pool>=3.1.0",
    "email-validator>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.6",
    "pgvector>=0.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
